refactor(channels): 重命名 ChannelPanel 和 SecretPanel 并将 useTenantConfigItems 移至 hooks
